LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/getresponseheader-error-state-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/nocors/folder.txt due to access control checks.

PASS XMLHttpRequest: getResponseHeader() in error state (failing cross-origin test)
